// src/app/model/user.ts
export class User {
    // id: any;
    // firstname: any;
    // lastname: any;
    // email: any;
    // password: any;
    // role: any;
    // logged: any;

    "id": 0;

    "firstname": "";

    "lastname": "";

    "email": "";

    "password": "";

    "role": "";

    "enabled": false;

    "accountNonExpired": false;

    "credentialsNonExpired": false;

    "authorities": [

        {

            "authority": ""

        }

    ];

    "username": "";

    "logged": '';

    "accountNonLocked": false;

}